Code.txt
This code uses the DHT library to interface with a DHT11 sensor and reads temperature and humidity values. Connect the DHT11 sensor to pin 2 on the Arduino board. Upload the code to your Arduino, and open the Serial Monitor to view the temperature and humidity readings.

Remember to install the DHT library in your Arduino IDE before uploading the code. You can do this by going to "Sketch" -> "Include Library" -> "Manage Libraries" and searching for "DHT".